+++ src/hit.cpp
@@ -3,7 +3,6 @@
 #include "hit.h"
 
 
-
 Hit::Hit(){
     this->t = 0;
     this->point = Vector3();
@@ -11,11 +10,9 @@ Hit::Hit(){
     this->color = Color();
 }
 
-Hit::Hit(float t, Vector3 point, Vector3 normal, Color color){
+Hit::Hit(double t, Vector3 point, Vector3 normal, Color color){
     this->t = t;
     this->point = point;
     this->normal = normal;
     this->color = color;
 }
-
-
