+++ src/vector2.cpp
@@ -6,7 +6,7 @@ Vector2::Vector2(){
     x = y = 0;
 }
 
-Vector2::Vector2(float x, float y){
+Vector2::Vector2(double x, double y){
     this->x = x;
     this->y = y;
 }
@@ -15,15 +15,17 @@ double Vector2::length(){
     return sqrt(x*x+y*y);
 }
 
-
 Vector2& Vector2::operator-=(Vector2 const& rhs){
     this->x = this->x - rhs.x;
     this->y = this->y - rhs.y;
     return *this;
 }
+
 Vector2 Vector2::operator-(Vector2 const& rhs){
     Vector2 vec;
     vec.x = this->x - rhs.x;
     vec.y = this->y - rhs.y;
     return vec;
 }
+
+
